requirements.txt
fastapi
uvicorn
jinja2
langchain
langchain-openai
python-dotenv
